+++ java/api/src/main/java/fr/ill/ics/cameo/coms/basic/Responder.java
@@ -67,6 +67,11 @@ public class Responder extends StateObject implements ICancelable {
 	@Override
 	public void init() throws InitException {
 
+		if (isReady()) {
+			// The object is already initialized.
+			return;
+		}
+		
 		// Set the key.
 		key = KEY + "-" + name;