+++ java/api/src/main/java/fr/ill/ics/cameo/coms/multi/ResponderRouter.java
@@ -60,6 +60,11 @@ public class ResponderRouter extends StateObject implements ICancelable {
 	@Override
 	public void init() throws InitException {
 
+		if (isReady()) {
+			// The object is already initialized.
+			return;
+		}
+		
 		// Set the key.
 		key = KEY + "-" + name;